ci: initial setup of workers in detox mock tests
