+++ apps/ledger-live-mobile/detox.config.js
@@ -13,6 +13,7 @@ module.exports = {
       setupTimeout: 500000,
     },
     retries: 0,
+    forwardEnv: true
   },
   logger: {
     level: process.env.DEBUG_DETOX ? "trace" : "info",
@@ -76,13 +77,41 @@ module.exports = {
         name: "iOS Simulator",
       },
     },
+    simulator2: {
+      type: "ios.simulator",
+      device: {
+        name: "iOS Simulator 2",
+      },
+    },
+    simulator3: {
+      type: "ios.simulator",
+      device: {
+        name: "iOS Simulator 3",
+      },
+    },
     emulator: {
       type: "android.emulator",
       device: {
         avdName: "Android_Emulator",
       },
       gpuMode: "swiftshader_indirect",
-      headless: process.env.CI ? true : false,
+      headless: !!process.env.CI,
+    },
+    emulator2: {
+      type: "android.emulator",
+      device: {
+        avdName: "Android_Emulator_2",
+      },
+      gpuMode: "swiftshader_indirect",
+      headless: !!process.env.CI,
+    },
+    emulator3: {
+      type: "android.emulator",
+      device: {
+        avdName: "Android_Emulator_3",
+      },
+      gpuMode: "swiftshader_indirect",
+      headless: !!process.env.CI,
     },
   },
   configurations: {
